requirements.txt
appdirs==1.4.4
attrs==24.2.0
autopep8==1.5.7
certifi==2024.8.30
chardet==3.0.4
click==8.0.1
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
git-bash==0.2
greenlet==1.1.0
gunicorn==20.1.0
html-testRunner==1.2.1
idna==2.10
importlib-metadata==4.0.1
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.22.0
mysqlclient==2.0.3
packaging==24.2
pathlib==1.0.1
pluggy==1.5.0
py==1.11.0
pycodestyle==2.7.0
PyJWT==2.9.0
pyprojroot==0.2.0
pytest==6.2.5
python-dotenv==1.0.1
PyYAML==5.4.1
requests==2.24.0
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.11
virtualenv==20.4.6
webargs==8.0.0
Werkzeug==2.0.1
wtforms==3.1.2
zipp==3.4.1
